Consolidate `readConfig` functions into one generic implementation
