+++ h/static/scripts/group-forms/config.ts
@@ -42,13 +42,4 @@ export type ConfigObject = {
   };
 };
 
-/** Return the frontend config from the page's <script class="js-config">. */
-export function readConfig(): ConfigObject {
-  try {
-    return JSON.parse(document.querySelector('.js-config')!.textContent!);
-  } catch {
-    throw new Error('Failed to parse frontend configuration');
-  }
-}
-
 export const Config = createContext<ConfigObject | null>(null);
