+++ h/static/scripts/group-forms/index.tsx
@@ -1,12 +1,13 @@
 import { render } from 'preact';
 
+import { readConfig } from '../forms-common/config';
 import AppRoot from './components/AppRoot';
-import { readConfig } from './config';
+import type { ConfigObject } from './config';
 
 function init() {
   const shadowHost = document.querySelector('#group-form')!;
   const shadowRoot = shadowHost.attachShadow({ mode: 'open' });
-  const config = readConfig();
+  const config = readConfig<ConfigObject>();
   render(<AppRoot config={config} />, shadowRoot);
 }
