+++ h/static/scripts/login-forms/config.ts
@@ -6,13 +6,4 @@ export type ConfigObject = {
   formData?: Record<string, string>;
 };
 
-/** Return the frontend config from the page's <script class="js-config">. */
-export function readConfig(): ConfigObject {
-  try {
-    return JSON.parse(document.querySelector('.js-config')!.textContent!);
-  } catch {
-    throw new Error('Failed to parse frontend configuration');
-  }
-}
-
 export const Config = createContext<ConfigObject | null>(null);
